Basics

Basics

Variables and Sets

The models scalability is relying on the usage of sets. The elements of the sets are extracted from the input data and scale the different variables. An overview of the sets is provided in the table. Depending on the models configuration the necessary sets are initialized.

name	description

lines	transmission lines connecting the nodes

nodes	spacial energy system nodes

tech	fossil and renewable generation as well as storage technologies

impact	impact categories like EUR or USD, CO 2 − eq., ...

account	fixed costs for installation and yearly expenses, variable costs

infrastruct	infrastructure status being either new or existing

sector	energy sector like electricity

time K	numeration of the representative periods

time T	numeration of the time intervals within a period

time T e	numeration of the time steps within a period

time I	numeration of the time invervals of the full input data periods

time I e	numeration of the time steps of the full input data periods

dir transmission	direction of the flow uniform with or opposite to the lines direction

An overview of the variables used in the CEP is provided in the table:

name	dimensions	unit	description

COST	[account,impact,tech]	EUR/USD, LCA-categories	Costs

CAP	[tech,infrastruct,node]	MW	Capacity

GEN	[sector,tech,t,k,node]	MW	Generation

SLACK	[sector,t,k,node]	MW	Power gap, not provided by installed CAP

LL	[sector]	MWh	LoastLoad Generation gap, not provided by installed CAP

LE	[impact]	LCA-categories	LoastEmission Amount of emissions that installed CAP crosses the Emission constraint

INTRASTOR	[sector, tech,t,k,node]	MWh	Storage level within a period

INTERSTOR	[sector,tech,i,node]	MWh	Storage level between periods of the full time series

FLOW	[sector,dir,tech,t,k,line]	MW	Flow over transmission line

TRANS	[tech,infrastruct,lines]	MW	maximum capacity of transmission lines

Data

The package provides data Capacity Expansion Data for:

name	nodes	lines	years	tech

GER-1	1 – germany as single node	none	2006-2016	Pv, wind, coal, oil, gas, bat-e, bat-in, bat-out, h2-e, h2-in, h2-out, trans

GER-18	18 – dena-zones within germany	49	2015	Pv, wind, coal, oil, gas, bat-e, bat-in, bat-out, h2-e, h2-in, h2-out, trans

CA-1	1 - california as single node	none	2016	Pv, wind, coal, oil, gas, bat-e, bat-in, bat-out, h2-e, h2-in, h2-out, trans

CA-14 ! currently not included !	14 – multiple nodes within CA and neighboring states	46	2016	Pv, wind, coal, oil, gas, bat-e, bat-in, bat-out, h2-e, h2-in, h2-out, trans

TX-1	1 – single node within Texas	none	2008	Pv, wind, coal, nuc, gas, bat-e, bat-in, bat-out

Units

Power - MW

Energy - MWh

lengths - km

Running the Capacity Expansion Problem

Note

The CEP model can be run with many configurations. The configurations themselves don't mess with each other though the provided input data must fulfill the ability to have e.g. lines in order for transmission to work.

An overview is provided in the following table:

description	unit	configuration	values	type	default value

enforce a CO2-limit	kg-CO2-eq./MW	co2_limit	>0	::Number	Inf

including existing infrastructure (no extra costs)	-	existing_infrastructure	true or false	::Bool	false

type of storage implementation	-	storage	"none", "simple" or "seasonal"	::String	"none"

allowing transmission	-	transmission	true or false	::Bool	false

fix. var and CEO to dispatch problem	-	fixeddesignvariables	design variables from design run or nothing	::OptVariables	nothing

allowing lost load (necessary for dispatch)	price/MWh	lostelload_cost	>1e6	::Number	Inf

allowing lost emission (necessary for dispatch)	price/kg_CO2-eq.	lostCO2emission_cost	>700	::Number	Inf

They can be applied in the following way:

CEP.run_opt — Function.

run_opt(ts_data::ClustData,opt_data::OptDataCEP,opt_config::Dict{String,Any};optimizer::DataType)

organizing the actual setup and run of the CEP-Problem

source

 run_opt(ts_data::ClustData,opt_data::OptDataCEP,opt_config::Dict{String,Any},fixed_design_variables::Dict{String,Any},optimizer::DataTyple;lost_el_load_cost::Number=Inf,lost_CO2_emission_cost::Number)

Wrapper function for type of optimization problem for the CEP-Problem (NOTE: identifier is the type of opt_data - in this case OptDataCEP - so identification as CEP problem) This problem runs the operational optimization problem only, with fixed design variables. provide the fixed design variables and the opt_config of the previous step (design run or another opterational run) what you can add to the opt_config:

lost_el_load_cost: Number indicating the lost load price/MWh (should be greater than 1e6), give Inf for none

lost_CO2_emission_cost: Number indicating the emission price/kg-CO2 (should be greater than 1e6), give Inf for none

give Inf for both lost_cost for no slack

source

 run_opt(ts_data::ClustData,opt_data::OptDataCEP,optimizer::DataTyple;descriptor::String="",co2_limit::Number=Inf,lost_el_load_cost::Number=Inf,lost_CO2_emission_cost::Number=Inf,existing_infrastructure::Bool=false,limit_infrastructure::Bool=false,storage::String="none",transmission::Bool=false,print_flag::Bool=true,print_level::Int64=0)

Wrapper function for type of optimization problem for the CEP-Problem (NOTE: identifier is the type of opt_data - in this case OptDataCEP - so identification as CEP problem) options to tweak the model are:

descritor: String with the name of this paricular model like "kmeans-10-co2-500"

co2_limit: A number limiting the kg.-CO2-eq./MWh (normally in a range from 5-1250 kg-CO2-eq/MWh), give Inf or no kw if unlimited

lost_el_load_cost: Number indicating the lost load price/MWh (should be greater than 1e6), give Inf for none

lost_CO2_emission_cost:

Number indicating the emission price/kg-CO2 (should be greater than 1e6), give Inf for none

give Inf for both lost_cost for no slack

existing_infrastructure: true or false to include or exclude existing infrastructure to the model

storage: String "none" for no storage or "simple" to include simple (only intra-day storage) or "seasonal" to include seasonal storage (inter-day)

source

Opt Result - A closer look

CEP.OptVariable — Type.

 OptVariable

-data::Array - includes the optimization variable output in form of an array -axes_names::Array{String,1}- includes the names of the different axes and is equivalent to the sets in the optimization formulation -axes::Tuple- includes the values of the different axes of the optimization variables -type::String` - defines the type of the variable being cv - cost variable - dv -design variable - ov - operating variable - sv - slack variable

source

CEP.OptResult — Type.

OptResult

source

Note

The model tracks how it is setup and which equations are used. This can help you to understand the models exact configuration without looking up the source code.

The information of the model setup can be checked out the following way:

result = run_opt(ts_clust_data.best_results,cep_data,optimizer;descriptor="Model Name")
result.opt_info["model"]