-

#Result Here, we describe the optimization result.

Types

The optimization results are provided in an OptResult struct:

Missing docstring.

Missing docstring for OptResult. Check Documenter's build log for details.

Equations

Note

The model tracks how it is setup and which equations are used. This can help you to understand the models exact configuration without looking up the source code.

The information of the model setup can be checked out the following way:

result = run_opt(ts_clust_data,cep_data,optimizer;descriptor="Model Name")
println.(result.opt_info["model"])

19-element Array{Nothing,1}:
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing
 nothing

Variables in Result

All variables are provided as dense OptVariable structs and can be indexed:

Missing docstring.

Missing docstring for OptVariable. Check Documenter's build log for details.

The variables can be of different type as explained in Variables. The different groups of variables can be extracted from the OptResult based on the variable_type:

CapacityExpansion.get_cep_variables — Function.

get_cep_variables(opt_result::OptResult, variable_type::String)

The variables can be of different type. The different groups of variables can be extracted from the OptResult based on the variable_type: Returns all variables which types match the String variable_type

source

CapacityExpansion.get_cep_slack_variables — Function.

get_cep_slack_variables(opt_result::OptResult)

Returns all slack variables in this opt_result mathing the type "sv"

source

CapacityExpansion.get_cep_design_variables — Function.

get_cep_design_variables(opt_result::OptResult)

Returns all design variables in this opt_result matching the type "dv" Additionally you can add capacity factors, which scale the design variables by multiplying it with the value in the Dict

source

The extraction of design variables is e.g. necessary for a Second stage operational validation step, which validates the energy system design on a different time series.

Printing information about the result

The function text_limit_emission allows to print an information about all the emission limits applied:

Missing docstring.

Missing docstring for text_limit_emission. Check Documenter's build log for details.