Preparing OptDataCEP

Preparing OptDataCEP

Loading Provided Data

load_cep_data_provided loads the non time-series dependent data for the CEP and can take the following regions:

GER: Germany

CA: California

TX: Texas

CEP.load_cep_data_provided — Function.

load_cep_data_provided(region::String)

Loading from .csv files in a the folder ../CEP/data/{region}/ Follow instructions preparing your own data: -region::String: name of state or region data belongs to -costs::OptVariable: costs[tech,node,year,account,impact] - annulized costs [USD in USD/MW_el, CO2 in kg-CO₂-eq./MW_el] -techs::OptVariable: techs[tech] - OptDataCEPTech -nodes::OptVariable: nodes[tech,node] - OptDataCEPNode -lines::OptVarible: lines[tech,line] - OptDataCEPLine for regions:

"GER_1": Germany 1 node

"GER_18": Germany 18 nodes

"CA_1": California 1 node

"CA_14": California 14 nodes

"TX_1": Texas 1 node

source

Loading Your Own Data

Use load_cep_data with data_path pointing to the folder with your cost, node, (line), and technology data.

CEP.load_cep_data — Function.

load_cep_data(data_path::String)

Loading from .csv files in a the folder /data_path/ Follow instructions for the CSV-Files: -region::String: name of state or region data belongs to -costs::OptVariable: costs[tech,node,year,account,impact] - annulized costs [USD in USD/MWel, CO2 in kg-CO₂-eq./MWel]-techs::OptVariable: techs[tech] - OptDataCEPTech - -nodes::OptVariable: nodes[tech,node] - OptDataCEPNode -lines::OptVarible`: lines[tech,line] - OptDataCEPLine

source

Data Types

CEP.OptVariable — Type.

 OptVariable

-data::Array - includes the optimization variable output in form of an array -axes_names::Array{String,1}- includes the names of the different axes and is equivalent to the sets in the optimization formulation -axes::Tuple- includes the values of the different axes of the optimization variables -type::String` - defines the type of the variable being cv - cost variable - dv -design variable - ov - operating variable - sv - slack variable

source

CEP.OptDataCEP — Type.

 OptDataCEP{region::String, costs::OptVariable, techs::OptVariable, nodes::OptVariable, lines::OptVariabl} <: OptData

region::String: name of state or region data belongs to

costs::OptVariable: costs[tech,node,year,account,impact] - Number

techs::OptVariable: techs[tech] - OptDataCEPTech

nodes::OptVariable: nodes[tech, node] - OptDataCEPNode

lines::OptVarible: lines[tech, line] - OptDataCEPLine

instead of USD you can also use your favorite currency like EUR

source

CEP.OptDataCEPTech — Type.

 OptDataCEPTech{name::String,categ::String,sector::String,eff::Number,time_series::String,lifetime::Number,financial_lifetime::Number,discount_rate::Number, annuityfactor::Number} <: OptData

name

categ: the category of this technology (is it storage, transmission or generation)

sector: sector of the technology (electricity or heat)

eff: efficiency of this technologies conversion [-]

time_series: time_series name for availability

lifetime: product lifetime [a]

financial_lifetime: financial time to break even [a]

discount_rate: discount rate for technology [a]

annuityfactor: annuity factor, important for cap-costs [-]

source

CEP.OptDataCEPNode — Type.

 OptDataCEPNode{name::String,value::Number,lat::Number,lon::Number} <: OptData

name

power_ex existing capacity [MW or MWh (tech_e)]

power_lim capacity limit [MW or MWh (tech_e)]

region

latlon hold geolocation information [°,°]

source

CEP.OptDataCEPLine — Type.

 OptDataCEPLine{name::String,node_start::String,node_end::String,reactance::Number,resistance::Number,power::Number,circuits::Int64,voltage::Number,length::Number} <: OptData

name

node_start Node where line starts

node_end Node where line ends

reactance

resistance [Ω]

power_ex: existing power limit [MW]

power_lim: limit power limit [MW]

circuits [-]

voltage [V]

length [km]

eff [-]

source

Examples

Example loading CEP Data

using CEP
state="GER_1"
# load ts-input-data
cep_data = load_cep_data_provided(state)
cep_data.costs

5-dimensional OptVariable{Number,5,...} of type fv with index sets:
    Dimension 1 - tech, ["pv", "wind", "coal", "oil", "gas", "bat_e", "bat_in", "bat_out", "h2_e", "h2_in", "h2_out", "trans"]
    Dimension 2 - node, ["germany"]
    Dimension 3 - year, [2015]
    Dimension 4 - account, ["cap_fix", "var"]
    Dimension 5 - impact, ["EUR", "CO2"]
And data, a 12×1×1×2×2 Array{Number,5}:
[:, :, 2015, "cap_fix", "EUR"] =
 136926.376
 168417.541
 153702.673
  38429.5834
  29964.5882
  25653.7085
   8961.67479
      0.0
     64.2184286
 479782.071
      0.0
     51.0061474

[:, :, 2015, "var", "EUR"] =
   0.0
   0.0
  10.708
 153.586
  86.564
   0.0
   0.0
   0.0
   0.0
   1.9
   0.0
   0.0

[:, :, 2015, "cap_fix", "CO2"] =
   160.233933
    36.75
    13.8595556
     8.25
     1.2334
  1658.35472
     0.0
     0.0
     0.0
 96666.6667
     0.0
     3.3275

[:, :, 2015, "var", "CO2"] =
    0.0
    0.0
 1217.49
  874.6
  728.71
    0.0
    0.0
    0.0
    0.0
    0.0
    0.0
    0.0

Example indexing OptVariables

Indexing is provided similar to Arrays:

cep_data.costs["pv","germany",2015,"cap_fix","EUR"]

136926.376

The axes are named and the axes can be called using the basic axes function and providing the sets name:

axes(cep_data.costs,"tech")

12-element Array{String,1}:
 "pv"
 "wind"
 "coal"
 "oil"
 "gas"
 "bat_e"
 "bat_in"
 "bat_out"
 "h2_e"
 "h2_in"
 "h2_out"
 "trans"